26/07/18: Ported to .NET Core, rewritten code and added features
DriveSP is ported to .NET Core for it to be compatible with Linix Based OS and MacOS, Code is re written to be more object oriented from before, and some features have been added. update command from Updater has been removed as it was similar to install command and served no purpose, detailed changes can be found inside the changes file in the repository.

Features added in this commit are:

In Publsiher,

- initialize without versionName

- setversion, set to target version by specifing VersionCode and VersionName, use --force option to set even if the target version exists on drive

- push now has force option to delete the existing version on drive with same, VersionCode that is being uploaded

- pushnow, increase version and push by just one command

- drop, delete a version which is on drive

- showversions, Display all the versions on drive

In Upater,

- install, VersionName or VersionCode now either can be used to target install version.

-checkupdate, check if new version is available.

- showversions, Display all the versions on drive.